rename new maintenance module
