+++ yt/fields/field_info_container.py
@@ -2,9 +2,9 @@ from numbers import Number as numeric_type
 
 import numpy as np
 
+from yt._maintenance.deprecation import issue_deprecation_warning
 from yt.funcs import mylog, only_on_root
 from yt.geometry.geometry_handler import is_curvilinear
-from yt.maintenance.deprecation import issue_deprecation_warning
 from yt.units.dimensions import dimensionless
 from yt.units.unit_object import Unit
 from yt.utilities.exceptions import YTFieldNotFound
