+++ yt/fields/species_fields.py
@@ -2,9 +2,9 @@ import re
 
 import numpy as np
 
+from yt._maintenance.deprecation import issue_deprecation_warning
 from yt.fields.field_detector import FieldDetector
 from yt.frontends.sph.data_structures import ParticleDataset
-from yt.maintenance.deprecation import issue_deprecation_warning
 from yt.utilities.chemical_formulas import ChemicalFormula
 from yt.utilities.physical_ratios import _primordial_mass_fraction
