+++ yt/visualization/eps_writer.py
@@ -2,8 +2,8 @@ import numpy as np
 import pyx
 from matplotlib import cm, pyplot as plt
 
+from yt._maintenance.deprecation import issue_deprecation_warning
 from yt.config import ytcfg
-from yt.maintenance.deprecation import issue_deprecation_warning
 from yt.units.unit_object import Unit
 from yt.units.yt_array import YTQuantity
 from yt.utilities.logger import ytLogger as mylog
